Do not pass down the grid ref as prop
