+++ webview/src/plots/components/customPlots/CustomPlotsGrid.tsx
@@ -1,10 +1,8 @@
-import React, { RefObject } from 'react'
+import React from 'react'
 import { PlotsSection } from 'dvc/src/plots/webview/contract'
-import { useObserveGridDimensions } from '../../hooks/useObserveGridDimensions'
 import { Grid } from '../Grid'
 
 interface CustomPlotsGridProps {
-  gridRef: RefObject<HTMLDivElement>
   nbItemsPerRow: number
   order: string[]
   parentDraggedOver: boolean
@@ -13,15 +11,12 @@ interface CustomPlotsGridProps {
 }
 
 export const CustomPlotsGrid: React.FC<CustomPlotsGridProps> = ({
-  gridRef,
   nbItemsPerRow,
   parentDraggedOver,
   order,
   setOrder,
   useVirtualizedGrid
 }) => {
-  useObserveGridDimensions(PlotsSection.CUSTOM_PLOTS, gridRef)
-
   return (
     <Grid
       setOrder={setOrder}
